Add test for failing get_columns_in_relation
Specifically, it fails when it's called on a model that was created in
the same run.

At the start of the run, the cache is populated, setting
Relation.information to be the string output of a DESCRIBE EXTENDED
query, which allows columns and metadata to be parsed.

However, when models are created, a Relation with information=None is
saved in the cache (since columns and metadata aren't returned from a
CREATE TABLE / VIEW statement). This means that an `expected string or
bytes-like object` error is raised when attempting to regex-parse None.