Fix get columns in relation (#197)
* Add test for failing get_columns_in_relation

Specifically, it fails when it's called on a model that was created in
the same run.

At the start of the run, the cache is populated, setting
Relation.information to be the string output of a DESCRIBE EXTENDED
query, which allows columns and metadata to be parsed.

However, when models are created, a Relation with information=None is
saved in the cache (since columns and metadata aren't returned from a
CREATE TABLE / VIEW statement). This means that an `expected string or
bytes-like object` error is raised when attempting to regex-parse None.

* Only parse cols from cache if there's information

If the `information` attribute is not yet set, we fall back on the
non-cached version to find column information. We could _also_ cache the
output of that query, but given that it wasn't cached originally, I
leave it as it is.

* Add get_columns_in_relation fix to CHANGELOG