some models for dbt-impala-example are not appearing in Hue when adapted to run with dbt-spark-livy adapter
Internal Ticket: https://jira.cloudera.com/projects/DBT/issues/DBT-124

Synopsis: When debuging the above ticket, the dbt logs indicated that there was an error in list_relations_without_caching indicating a NoneType was passed to the macro.
Further it was observed that when an actual error by SQL executing in the warehouse was thrown, the error was siliently ignored.
This PR addresses this issue.

Testplan:
Clone and sucessfully execute dbt seed and dbt run for the project https://github.com/TapasSenapati/dbt-spark-example